requirements.txt
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-limiter==0.1.5
frozenlist==1.4.0
google-auth==2.23.0
greenlet==2.0.2
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
packaging==23.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
unpaddedbase64==2.1.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
